Brand 1st Commit

+++ laravel/app/Http/Controllers/PostController.php
@@ -62,13 +62,14 @@ class PostController extends Controller
         }
         $user = Auth::user();
         $like = $user->likes()->where('post_id', $post_id)->first();
+        if($like){
             $already_like = $like->like;
             $update = true;
             if ($already_like == $is_like) {
                 $like->delete();
                 return null;
             }
-         else {
+        } else {
             $like = new Like();
         }
         $like->like = $is_like;
